ref: 部屋の存在チェックを別のコンポーネントとして切り出す
